Add files via upload
This Project aims at creating a practical knowledge, experience and exposure for the students by giving an opportunity to know, observe, and learn the  vision, objectives, and functioning of an Event management system and the activities in that organization. 
Event  Management System (EMS) was introduced to solve the complications coming from managing all the paperwork of every event associated with peoples and their families. EMS provides the ability to manage all the paperwork in one place, reducing the work of staff in arranging and analyzing the paperwork of the managing the event . EMS does many works like: